refactor: clean up object detection visualization and utility functions
